# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Qutlas API
  version: 1.0.0
  description: Browser-first CAD/CAM distributed micro-manufacturing platform
  contact:
    name: Qutlas Support
    url: https://qutlas.com

servers:
  - url: https://api.qutlas.com/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development

tags:
  - name: Auth
    description: Authentication and authorization
  - name: Assets
    description: CAD file upload and management
  - name: Catalog
    description: Parametric parts catalog
  - name: Jobs
    description: Manufacturing job lifecycle
  - name: Hubs
    description: Hub management and onboarding
  - name: AI
    description: AI-powered manufacturability assessment
  - name: Geometry
    description: Internal geometry operations (OCCT)

paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Jane Doe"
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                company:
                  type: string
                role:
                  type: string
                  enum: [designer, manufacturer, admin]
                invite_code:
                  type: string
                  nullable: true
              required: [name, email, password, role]
      responses:
        "201":
          description: User created; email verification link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  verification_sent:
                    type: boolean
        "400":
          description: Validation error
        "409":
          description: Email already registered

  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
        "401":
          description: Invalid credentials

  /auth/oidc/callback:
    post:
      tags: [Auth]
      summary: OIDC provider callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                state:
                  type: string
              required: [code, state]
      responses:
        "200":
          description: Auth successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user_id:
                    type: string

  /auth/hub/register:
    post:
      tags: [Auth]
      summary: Hub certification application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                company_address:
                  type: string
                machines:
                  type: array
                  items:
                    type: object
                    properties:
                      machine_type:
                        type: string
                        enum: [cnc_mill, cnc_lathe, laser, waterjet, 3d_printer, robot_arm]
                      model:
                        type: string
                      year:
                        type: integer
                      capabilities:
                        type: array
                        items:
                          type: string
                    required: [machine_type, model, year]
                insurance_docs:
                  type: array
                  items:
                    type: object
                    properties:
                      document_type:
                        type: string
                        enum: [general_liability, workers_comp, product_liability]
                      file_url:
                        type: string
                    required: [document_type, file_url]
              required: [company_name, company_address, machines, insurance_docs]
      responses:
        "202":
          description: Hub application submitted for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  application_id:
                    type: string
                  status:
                    type: string
                    enum: [submitted, under_review, approved, rejected]
        "400":
          description: Validation error

  /assets/upload:
    post:
      tags: [Assets]
      summary: Upload CAD file (STEP, IGES, STL)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CAD file (STEP, IGES, STL)
                metadata:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
              required: [file]
      responses:
        "201":
          description: File uploaded and validated
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id:
                    type: string
                  qdf_id:
                    type: string
                  filename:
                    type: string
                  validation_status:
                    type: string
                    enum: [valid, warnings, errors]
                  validation_issues:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        message:
                          type: string
                        severity:
                          type: string
                          enum: [info, warning, error]
                  thumbnail_url:
                    type: string
                  bounding_box:
                    type: object
                    properties:
                      min:
                        type: array
                        items: [number, number, number]
                      max:
                        type: array
                        items: [number, number, number]
        "400":
          description: Invalid file format or corrupted
        "413":
          description: File too large (max 500 MB)

  /catalog:
    post:
      tags: [Catalog]
      summary: Create catalog item (user-supplied or supplier part)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogItem"
      responses:
        "201":
          description: Catalog item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
        "400":
          description: Validation error

    get:
      tags: [Catalog]
      summary: List catalog items with search and filter
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (title, description)
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Catalog items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CatalogItem"
                  total:
                    type: integer

  /catalog/{id}:
    get:
      tags: [Catalog]
      summary: Get catalog item detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Catalog item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
        "404":
          description: Not found

  /catalog/{id}/hubs:
    get:
      tags: [Catalog]
      summary: Find hubs capable of manufacturing item variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: variant_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of matching hubs
          content:
            application/json:
              schema:
                type: object
                properties:
                  hubs:
                    type: array
                    items:
                      $ref: "#/components/schemas/HubSummary"

  /jobs/create:
    post:
      tags: [Jobs]
      summary: Create manufacturing job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                catalog_item_id:
                  type: string
                variant_id:
                  type: string
                parameters:
                  type: object
                  description: Variant-specific numeric parameters
                hub_id:
                  type: string
                customer_id:
                  type: string
                payment_method:
                  type: string
                  description: Stripe payment method ID
              required: [catalog_item_id, variant_id, hub_id, customer_id]
      responses:
        "201":
          description: Job created and routed to hub
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [created, hub_accepted, in_progress, completed, failed]
                  hub_id:
                    type: string
                  estimated_completion:
                    type: string
                    format: date-time
                  cost_estimate:
                    type: number
                  cost_currency:
                    type: string
        "400":
          description: Validation error or hub unavailable

  /jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get job status and details
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{job_id}/stream:
    get:
      tags: [Jobs]
      summary: Stream toolpath and status updates (SSE or WebSocket)
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server-sent event stream
          content:
            text/event-stream:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    enum: [segment, status, error]
                  data:
                    type: object

  /ai/v1/assess:
    post:
      tags: [AI]
      summary: Assess manufacturability of asset + variant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id:
                  type: string
                variant_id:
                  type: string
                  nullable: true
                parameters:
                  type: object
                  description: Variant parameter overrides
              required: [asset_id]
      responses:
        "200":
          description: Manufacturability assessment
          content:
            application/json:
              schema:
                type: object
                properties:
                  manufacturability_score:
                    type: number
                    minimum: 0
                    maximum: 1
                  issues:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                          example: "wall_too_thin"
                        message:
                          type: string
                        severity:
                          type: string
                          enum: [info, warning, error]
                        suggested_fix:
                          type: object
                          properties:
                            action:
                              type: string
                            parameter:
                              type: string
                            value:
                              type: number
                  processing_time_ms:
                    type: integer

  /internal/geometry/boolean:
    post:
      tags: [Geometry]
      summary: Perform robust boolean operation on two B-Rep bodies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operation:
                  type: string
                  enum: [union, difference, intersection]
                body_a_id:
                  type: string
                body_b_id:
                  type: string
              required: [operation, body_a_id, body_b_id]
      responses:
        "202":
          description: Boolean operation queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string

  /internal/geometry/ingest:
    post:
      tags: [Geometry]
      summary: Ingest raw STEP/IGES file and extract parametric metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Ingestion queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id:
                    type: string
                  job_id:
                    type: string

components:
  schemas:
    CatalogItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        qdf_ref:
          type: string
          description: Reference to QDF or STEP asset
        supplier:
          type: object
          properties:
            name:
              type: string
            url:
              type: string
        variants:
          type: array
          items:
            type: object
            properties:
              variant_id:
                type: string
              parameters:
                type: object
              material:
                type: string
              price_base:
                type: number
              price_currency:
                type: string
                enum: [CAD, USD, EUR]
              lead_time_days:
                type: integer
              manufacturing_methods:
                type: array
                items:
                  type: string
              hub_tags:
                type: array
                items:
                  type: string
            required: [variant_id, parameters]
        ai_metrics:
          type: object
          properties:
            manufacturability:
              type: number
        assets:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
        visibility:
          type: string
          enum: [public, private, hub-only]
        created_at:
          type: string
          format: date-time
      required: [id, title, qdf_ref, variants]

    HubSummary:
      type: object
      properties:
        hub_id:
          type: string
        name:
          type: string
        location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
            city:
              type: string
            country:
              type: string
        machines:
          type: array
          items:
            type: string
        certification_level:
          type: string
          enum: [basic, verified, premium]
        average_rating:
          type: number
        current_load:
          type: number
          description: Queue load 0-1
        estimated_price:
          type: number
        estimated_delivery_days:
          type: integer

    Job:
      type: object
      properties:
        job_id:
          type: string
        catalog_item_id:
          type: string
        variant_id:
          type: string
        hub_id:
          type: string
        status:
          type: string
          enum: [created, hub_accepted, in_progress, completed, failed]
        customer_id:
          type: string
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        cost_estimate:
          type: number
        cost_final:
          type: number
        first_pass_yield:
          type: number
          description: "0-1; null if not completed"